Se realiza la logica de funcionamiento y se ajustan las pantallas

+++ dispenser_prj_ise/dependencies/modules/bottle_level_control/bottle_level_control.h
@@ -15,7 +15,7 @@
 
 
 
-uint8_t getLiquidLevel();
+float getLiquidLevel();
 void liquidControlInit();
 
 //=====[#include guards - end]=================================================
